@@ -1,124 +1,110 @@
-var tape = require("tape"),
-    array = require("d3-array"),
-    d3 = require("../");
+import assert from "assert";
+import * as d3 from "../src/index.js";
 
-tape("a sphere contains any point", function(test) {
-  test.equal(d3.geoContains({type: "Sphere"}, [0, 0]), true);
-  test.end();
+it("a sphere contains any point", () => {
+  assert.strictEqual(d3.geoContains({type: "Sphere"}, [0, 0]), true);
 });
 
-tape("a point contains itself (and not some other point)", function(test) {
-  test.equal(d3.geoContains({type: "Point", coordinates: [0, 0]}, [0, 0]), true);
-  test.equal(d3.geoContains({type: "Point", coordinates: [1, 2]}, [1, 2]), true);
-  test.equal(d3.geoContains({type: "Point", coordinates: [0, 0]}, [0, 1]), false);
-  test.equal(d3.geoContains({type: "Point", coordinates: [1, 1]}, [1, 0]), false);
-  test.end();
+it("a point contains itself (and not some other point)", () => {
+  assert.strictEqual(d3.geoContains({type: "Point", coordinates: [0, 0]}, [0, 0]), true);
+  assert.strictEqual(d3.geoContains({type: "Point", coordinates: [1, 2]}, [1, 2]), true);
+  assert.strictEqual(d3.geoContains({type: "Point", coordinates: [0, 0]}, [0, 1]), false);
+  assert.strictEqual(d3.geoContains({type: "Point", coordinates: [1, 1]}, [1, 0]), false);
 });
 
-tape("a MultiPoint contains any of its points", function(test) {
-  test.equal(d3.geoContains({type: "MultiPoint", coordinates: [[0, 0], [1,2]]}, [0, 0]), true);
-  test.equal(d3.geoContains({type: "MultiPoint", coordinates: [[0, 0], [1,2]]}, [1, 2]), true);
-  test.equal(d3.geoContains({type: "MultiPoint", coordinates: [[0, 0], [1,2]]}, [1, 3]), false);
-  test.end();
+it("a MultiPoint contains any of its points", () => {
+  assert.strictEqual(d3.geoContains({type: "MultiPoint", coordinates: [[0, 0], [1,2]]}, [0, 0]), true);
+  assert.strictEqual(d3.geoContains({type: "MultiPoint", coordinates: [[0, 0], [1,2]]}, [1, 2]), true);
+  assert.strictEqual(d3.geoContains({type: "MultiPoint", coordinates: [[0, 0], [1,2]]}, [1, 3]), false);
 });
 
-tape("a LineString contains any point on the Great Circle path", function(test) {
-  test.equal(d3.geoContains({type: "LineString", coordinates: [[0, 0], [1,2]]}, [0, 0]), true);
-  test.equal(d3.geoContains({type: "LineString", coordinates: [[0, 0], [1,2]]}, [1, 2]), true);
-  test.equal(d3.geoContains({type: "LineString", coordinates: [[0, 0], [1,2]]}, d3.geoInterpolate([0, 0], [1,2])(0.3)), true);
-  test.equal(d3.geoContains({type: "LineString", coordinates: [[0, 0], [1,2]]}, d3.geoInterpolate([0, 0], [1,2])(1.3)), false);
-  test.equal(d3.geoContains({type: "LineString", coordinates: [[0, 0], [1,2]]}, d3.geoInterpolate([0, 0], [1,2])(-0.3)), false);
-  test.end();
+it("a LineString contains any point on the Great Circle path", () => {
+  assert.strictEqual(d3.geoContains({type: "LineString", coordinates: [[0, 0], [1,2]]}, [0, 0]), true);
+  assert.strictEqual(d3.geoContains({type: "LineString", coordinates: [[0, 0], [1,2]]}, [1, 2]), true);
+  assert.strictEqual(d3.geoContains({type: "LineString", coordinates: [[0, 0], [1,2]]}, d3.geoInterpolate([0, 0], [1,2])(0.3)), true);
+  assert.strictEqual(d3.geoContains({type: "LineString", coordinates: [[0, 0], [1,2]]}, d3.geoInterpolate([0, 0], [1,2])(1.3)), false);
+  assert.strictEqual(d3.geoContains({type: "LineString", coordinates: [[0, 0], [1,2]]}, d3.geoInterpolate([0, 0], [1,2])(-0.3)), false);
 });
 
-tape("a LineString with 2+ points contains those points", function(test) {
-  var points = [[0, 0], [1,2], [3, 4], [5, 6]];
-  var feature = {type: "LineString", coordinates: points};
+it("a LineString with 2+ points contains those points", () => {
+  const points = [[0, 0], [1,2], [3, 4], [5, 6]];
+  const feature = {type: "LineString", coordinates: points};
   points.forEach(point => {
-    test.equal(d3.geoContains(feature, point), true);
+    assert.strictEqual(d3.geoContains(feature, point), true);
   });
-  test.end();
 });
 
-tape("a LineString contains epsilon-distant points", function(test) {
-  var epsilon = 1e-6;
-  var line = [[0, 0], [0, 10], [10, 10], [10, 0]];
-  var points = [[0, 5], [epsilon * 1, 5], [0, epsilon], [epsilon * 1, epsilon]];
+it("a LineString contains epsilon-distant points", () => {
+  const epsilon = 1e-6;
+  const line = [[0, 0], [0, 10], [10, 10], [10, 0]];
+  const points = [[0, 5], [epsilon * 1, 5], [0, epsilon], [epsilon * 1, epsilon]];
   points.forEach(point => {
-    test.true(d3.geoContains({type:"LineString", coordinates: line}, point));
+    assert(d3.geoContains({type:"LineString", coordinates: line}, point));
   });
-  test.end();
 });
 
-tape("a LineString does not contain 10*epsilon-distant points", function(test) {
-  var epsilon = 1e-6;
-  var line = [[0, 0], [0, 10], [10, 10], [10, 0]];
-  var points = [[epsilon * 10, 5], [epsilon * 10, epsilon]];
+it("a LineString does not contain 10*epsilon-distant points", () => {
+  const epsilon = 1e-6;
+  const line = [[0, 0], [0, 10], [10, 10], [10, 0]];
+  const points = [[epsilon * 10, 5], [epsilon * 10, epsilon]];
   points.forEach(point => {
-    test.false(d3.geoContains({type:"LineString", coordinates: line}, point));
+    assert(!d3.geoContains({type:"LineString", coordinates: line}, point));
   });
-  test.end();
 });
 
-tape("a MultiLineString contains any point on one of its components", function(test) {
-  test.equal(d3.geoContains({type: "MultiLineString", coordinates: [[[0, 0], [1,2]], [[2, 3], [4,5]]]}, [2, 3]), true);
-  test.equal(d3.geoContains({type: "MultiLineString", coordinates: [[[0, 0], [1,2]], [[2, 3], [4,5]]]}, [5, 6]), false);
-  test.end();
+it("a MultiLineString contains any point on one of its components", () => {
+  assert.strictEqual(d3.geoContains({type: "MultiLineString", coordinates: [[[0, 0], [1,2]], [[2, 3], [4,5]]]}, [2, 3]), true);
+  assert.strictEqual(d3.geoContains({type: "MultiLineString", coordinates: [[[0, 0], [1,2]], [[2, 3], [4,5]]]}, [5, 6]), false);
 });
 
-tape("a Polygon contains a point", function(test) {
-  var polygon = d3.geoCircle().radius(60)();
-  test.equal(d3.geoContains(polygon, [1, 1]), true);
-  test.equal(d3.geoContains(polygon, [-180, 0]), false);
-  test.end();
+it("a Polygon contains a point", () => {
+  const polygon = d3.geoCircle().radius(60)();
+  assert.strictEqual(d3.geoContains(polygon, [1, 1]), true);
+  assert.strictEqual(d3.geoContains(polygon, [-180, 0]), false);
 });
 
-tape("a Polygon with a hole doesn't contain a point", function(test) {
-  var outer = d3.geoCircle().radius(60)().coordinates[0],
+it("a Polygon with a hole doesn't contain a point", () => {
+  const outer = d3.geoCircle().radius(60)().coordinates[0],
       inner = d3.geoCircle().radius(3)().coordinates[0],
       polygon = {type:"Polygon", coordinates: [outer, inner]};
-  test.equal(d3.geoContains(polygon, [1, 1]), false);
-  test.equal(d3.geoContains(polygon, [5, 0]), true);
-  test.equal(d3.geoContains(polygon, [65, 0]), false);
-  test.end();
+  assert.strictEqual(d3.geoContains(polygon, [1, 1]), false);
+  assert.strictEqual(d3.geoContains(polygon, [5, 0]), true);
+  assert.strictEqual(d3.geoContains(polygon, [65, 0]), false);
 });
 
-tape("a MultiPolygon contains a point", function(test) {
-  var p1 = d3.geoCircle().radius(6)().coordinates,
+it("a MultiPolygon contains a point", () => {
+  const p1 = d3.geoCircle().radius(6)().coordinates,
       p2 = d3.geoCircle().radius(6).center([90,0])().coordinates,
       polygon = {type:"MultiPolygon", coordinates: [p1, p2]};
-  test.equal(d3.geoContains(polygon, [1, 0]), true);
-  test.equal(d3.geoContains(polygon, [90, 1]), true);
-  test.equal(d3.geoContains(polygon, [90, 45]), false);
-  test.end();
+  assert.strictEqual(d3.geoContains(polygon, [1, 0]), true);
+  assert.strictEqual(d3.geoContains(polygon, [90, 1]), true);
+  assert.strictEqual(d3.geoContains(polygon, [90, 45]), false);
 });
 
-tape("a GeometryCollection contains a point", function(test) {
-  var collection = {
+it("a GeometryCollection contains a point", () => {
+  const collection = {
     type: "GeometryCollection", geometries: [
       {type: "GeometryCollection", geometries: [{type: "LineString", coordinates: [[-45, 0], [0, 0]]}]},
       {type: "LineString", coordinates: [[0, 0], [45, 0]]}
     ]
   };
-  test.equal(d3.geoContains(collection, [-45, 0]), true);
-  test.equal(d3.geoContains(collection, [45, 0]), true);
-  test.equal(d3.geoContains(collection, [12, 25]), false);
-  test.end();
+  assert.strictEqual(d3.geoContains(collection, [-45, 0]), true);
+  assert.strictEqual(d3.geoContains(collection, [45, 0]), true);
+  assert.strictEqual(d3.geoContains(collection, [12, 25]), false);
 });
 
-tape("a Feature contains a point", function(test) {
-  var feature = {
+it("a Feature contains a point", () => {
+  const feature = {
     type: "Feature", geometry: {
       type: "LineString", coordinates: [[0, 0], [45, 0]]
     }
   };
-  test.equal(d3.geoContains(feature, [45, 0]), true);
-  test.equal(d3.geoContains(feature, [12, 25]), false);
-  test.end();
+  assert.strictEqual(d3.geoContains(feature, [45, 0]), true);
+  assert.strictEqual(d3.geoContains(feature, [12, 25]), false);
 });
 
-tape("a FeatureCollection contains a point", function(test) {
-  var feature1 = {
+it("a FeatureCollection contains a point", () => {
+  const feature1 = {
     type: "Feature", geometry: {
       type: "LineString", coordinates: [[0, 0], [45, 0]]
     }
@@ -132,14 +118,12 @@ tape("a FeatureCollection contains a point", function(test) {
     type: "FeatureCollection",
     features: [ feature1, feature2 ]
   };
-  test.equal(d3.geoContains(featureCollection, [45, 0]), true);
-  test.equal(d3.geoContains(featureCollection, [-45, 0]), true);
-  test.equal(d3.geoContains(featureCollection, [12, 25]), false);
-  test.end();
+  assert.strictEqual(d3.geoContains(featureCollection, [45, 0]), true);
+  assert.strictEqual(d3.geoContains(featureCollection, [-45, 0]), true);
+  assert.strictEqual(d3.geoContains(featureCollection, [12, 25]), false);
 });
 
-tape("null contains nothing", function(test) {
-  test.equal(d3.geoContains(null, [0, 0]), false);
-  test.end();
+it("null contains nothing", () => {
+  assert.strictEqual(d3.geoContains(null, [0, 0]), false);
 });
 
