@@ -1,61 +1,50 @@
-var tape = require("tape"),
-    d3 = require("../");
+import * as d3 from "../src/index.js";
+import {assertInDelta} from "./asserts.js";
 
-require("./inDelta");
-
-tape("geoLength(Point) returns zero", function(test) {
-  test.inDelta(d3.geoLength({type: "Point", coordinates: [0, 0]}), 0, 1e-6);
-  test.end();
+it("geoLength(Point) returns zero", () => {
+  assertInDelta(d3.geoLength({type: "Point", coordinates: [0, 0]}), 0, 1e-6);
 });
 
-tape("geoLength(MultiPoint) returns zero", function(test) {
-  test.inDelta(d3.geoLength({type: "MultiPoint", coordinates: [[0, 1], [2, 3]]}), 0, 1e-6);
-  test.end();
+it("geoLength(MultiPoint) returns zero", () => {
+  assertInDelta(d3.geoLength({type: "MultiPoint", coordinates: [[0, 1], [2, 3]]}), 0, 1e-6);
 });
 
-tape("geoLength(LineString) returns the sum of its great-arc segments", function(test) {
-  test.inDelta(d3.geoLength({type: "LineString", coordinates: [[-45, 0], [45, 0]]}), Math.PI / 2, 1e-6);
-  test.inDelta(d3.geoLength({type: "LineString", coordinates: [[-45, 0], [-30, 0], [-15, 0], [0, 0]]}), Math.PI / 4, 1e-6);
-  test.end();
+it("geoLength(LineString) returns the sum of its great-arc segments", () => {
+  assertInDelta(d3.geoLength({type: "LineString", coordinates: [[-45, 0], [45, 0]]}), Math.PI / 2, 1e-6);
+  assertInDelta(d3.geoLength({type: "LineString", coordinates: [[-45, 0], [-30, 0], [-15, 0], [0, 0]]}), Math.PI / 4, 1e-6);
 });
 
-tape("geoLength(MultiLineString) returns the sum of its great-arc segments", function(test) {
-  test.inDelta(d3.geoLength({type: "MultiLineString", coordinates: [[[-45, 0], [-30, 0]], [[-15, 0], [0, 0]]]}), Math.PI / 6, 1e-6);
-  test.end();
+it("geoLength(MultiLineString) returns the sum of its great-arc segments", () => {
+  assertInDelta(d3.geoLength({type: "MultiLineString", coordinates: [[[-45, 0], [-30, 0]], [[-15, 0], [0, 0]]]}), Math.PI / 6, 1e-6);
 });
 
-tape("geoLength(Polygon) returns the length of its perimeter", function(test) {
-  test.inDelta(d3.geoLength({type: "Polygon", coordinates: [[[0, 0], [3, 0], [3, 3], [0, 3], [0, 0]]]}), 0.157008, 1e-6);
-  test.end();
+it("geoLength(Polygon) returns the length of its perimeter", () => {
+  assertInDelta(d3.geoLength({type: "Polygon", coordinates: [[[0, 0], [3, 0], [3, 3], [0, 3], [0, 0]]]}), 0.157008, 1e-6);
 });
 
-tape("geoLength(Polygon) returns the length of its perimeter, including holes", function(test) {
-  test.inDelta(d3.geoLength({type: "Polygon", coordinates: [[[0, 0], [3, 0], [3, 3], [0, 3], [0, 0]], [[1, 1], [2, 1], [2, 2], [1, 2], [1, 1]]]}), 0.209354, 1e-6);
-  test.end();
+it("geoLength(Polygon) returns the length of its perimeter, including holes", () => {
+  assertInDelta(d3.geoLength({type: "Polygon", coordinates: [[[0, 0], [3, 0], [3, 3], [0, 3], [0, 0]], [[1, 1], [2, 1], [2, 2], [1, 2], [1, 1]]]}), 0.209354, 1e-6);
 });
 
-tape("geoLength(MultiPolygon) returns the summed length of the perimeters", function(test) {
-  test.inDelta(d3.geoLength({type: "MultiPolygon", coordinates: [[[[0, 0], [3, 0], [3, 3], [0, 3], [0, 0]]]]}), 0.157008, 1e-6);
-  test.inDelta(d3.geoLength({type: "MultiPolygon", coordinates: [[[[0, 0], [3, 0], [3, 3], [0, 3], [0, 0]]], [[[1, 1], [2, 1], [2, 2], [1, 2], [1, 1]]]]}), 0.209354, 1e-6);
-  test.end();
+it("geoLength(MultiPolygon) returns the summed length of the perimeters", () => {
+  assertInDelta(d3.geoLength({type: "MultiPolygon", coordinates: [[[[0, 0], [3, 0], [3, 3], [0, 3], [0, 0]]]]}), 0.157008, 1e-6);
+  assertInDelta(d3.geoLength({type: "MultiPolygon", coordinates: [[[[0, 0], [3, 0], [3, 3], [0, 3], [0, 0]]], [[[1, 1], [2, 1], [2, 2], [1, 2], [1, 1]]]]}), 0.209354, 1e-6);
 });
 
-tape("geoLength(FeatureCollection) returns the sum of its features’ lengths", function(test) {
-  test.inDelta(d3.geoLength({
+it("geoLength(FeatureCollection) returns the sum of its features’ lengths", () => {
+  assertInDelta(d3.geoLength({
     type: "FeatureCollection", features: [
       {type: "Feature", geometry: {type: "LineString", coordinates: [[-45, 0], [0, 0]]}},
       {type: "Feature", geometry: {type: "LineString", coordinates: [[0, 0], [45, 0]]}}
     ]
   }), Math.PI / 2, 1e-6);
-  test.end();
 });
 
-tape("geoLength(GeometryCollection) returns the sum of its geometries’ lengths", function(test) {
-  test.inDelta(d3.geoLength({
+it("geoLength(GeometryCollection) returns the sum of its geometries’ lengths", () => {
+  assertInDelta(d3.geoLength({
     type: "GeometryCollection", geometries: [
       {type: "GeometryCollection", geometries: [{type: "LineString", coordinates: [[-45, 0], [0, 0]]}]},
       {type: "LineString", coordinates: [[0, 0], [45, 0]]}
     ]
   }), Math.PI / 2, 1e-6);
-  test.end();
 });
