@@ -1,35 +1,32 @@
-var tape = require("tape"),
-    d3_geo = require("../../");
+import assert from "assert";
+import * as d3 from "../../src/index.js";
 
-var equirectangular = d3_geo.geoEquirectangular()
+const equirectangular = d3.geoEquirectangular()
     .scale(900 / Math.PI)
     .precision(0);
 
 function testArea(projection, object) {
-  return d3_geo.geoPath()
+  return d3.geoPath()
       .projection(projection)
       .area(object);
 }
 
-tape("geoPath.area(…) of a polygon with no holes", function(test) {
-  test.equal(testArea(equirectangular, {
+it("geoPath.area(…) of a polygon with no holes", () => {
+  assert.strictEqual(testArea(equirectangular, {
     type: "Polygon",
     coordinates: [[[100, 0], [100, 1], [101, 1], [101, 0], [100, 0]]]
   }), 25);
-  test.end();
 });
 
-tape("geoPath.area(…) of a polygon with holes", function(test) {
-  test.equal(testArea(equirectangular, {
+it("geoPath.area(…) of a polygon with holes", () => {
+  assert.strictEqual(testArea(equirectangular, {
     type: "Polygon",
     coordinates: [[[100, 0], [100, 1], [101, 1], [101, 0], [100, 0]], [[100.2, 0.2], [100.8, 0.2], [100.8, 0.8], [100.2, 0.8], [100.2, 0.2]]]
   }), 16);
-  test.end();
 });
 
-tape("geoPath.area(…) of a sphere", function(test) {
-  test.equal(testArea(equirectangular, {
+it("geoPath.area(…) of a sphere", () => {
+  assert.strictEqual(testArea(equirectangular, {
     type: "Sphere",
   }), 1620000);
-  test.end();
 });
