@@ -1,17 +1,14 @@
-var tape = require("tape"),
-    d3 = require("../../");
+import assert from "assert";
+import * as d3 from "../../src/index.js";
 
-require("../pathEqual");
-
-tape("azimuthal projections don't crash on the antipode", function(test) {
+it("azimuthal projections don't crash on the antipode", () => {
   for (const p of [
     d3.geoAzimuthalEqualArea()([180, 0]),
     d3.geoAzimuthalEqualArea()([-180, 0]),
     d3.geoAzimuthalEquidistant()([180, 0])
   ]) {
-    test.assert(Math.abs(p[0]) < Infinity);
-    test.assert(Math.abs(p[1]) < Infinity);
-  };
-  test.end();
+    assert(Math.abs(p[0]) < Infinity);
+    assert(Math.abs(p[1]) < Infinity);
+  }
 });
 
