@@ -1,45 +1,39 @@
-var tape = require("tape"),
-    d3 = require("../../");
+import assert from "assert";
+import * as d3 from "../../src/index.js";
+import {assertPathEqual} from "../asserts.js";
 
-require("../pathEqual");
-
-tape("mercator.clipExtent(null) sets the default automatic clip extent", function(test) {
-  var projection = d3.geoMercator().translate([0, 0]).scale(1).clipExtent(null).precision(0);
-  test.pathEqual(d3.geoPath(projection)({type: "Sphere"}), "M3.141593,-3.141593L3.141593,0L3.141593,3.141593L3.141593,3.141593L-3.141593,3.141593L-3.141593,3.141593L-3.141593,0L-3.141593,-3.141593L-3.141593,-3.141593L3.141593,-3.141593Z");
-  test.equal(projection.clipExtent(), null);
-  test.end();
+it("mercator.clipExtent(null) sets the default automatic clip extent", () => {
+  const projection = d3.geoMercator().translate([0, 0]).scale(1).clipExtent(null).precision(0);
+  assertPathEqual(d3.geoPath(projection)({type: "Sphere"}), "M3.141593,-3.141593L3.141593,0L3.141593,3.141593L3.141593,3.141593L-3.141593,3.141593L-3.141593,3.141593L-3.141593,0L-3.141593,-3.141593L-3.141593,-3.141593L3.141593,-3.141593Z");
+  assert.strictEqual(projection.clipExtent(), null);
 });
 
-tape("mercator.center(center) sets the correct automatic clip extent", function(test) {
-  var projection = d3.geoMercator().translate([0, 0]).scale(1).center([10, 10]).precision(0);
-  test.pathEqual(d3.geoPath(projection)({type: "Sphere"}), "M2.967060,-2.966167L2.967060,0.175426L2.967060,3.317018L2.967060,3.317018L-3.316126,3.317018L-3.316126,3.317019L-3.316126,0.175426L-3.316126,-2.966167L-3.316126,-2.966167L2.967060,-2.966167Z");
-  test.equal(projection.clipExtent(), null);
-  test.end();
+it("mercator.center(center) sets the correct automatic clip extent", () => {
+  const projection = d3.geoMercator().translate([0, 0]).scale(1).center([10, 10]).precision(0);
+  assertPathEqual(d3.geoPath(projection)({type: "Sphere"}), "M2.967060,-2.966167L2.967060,0.175426L2.967060,3.317018L2.967060,3.317018L-3.316126,3.317018L-3.316126,3.317019L-3.316126,0.175426L-3.316126,-2.966167L-3.316126,-2.966167L2.967060,-2.966167Z");
+  assert.strictEqual(projection.clipExtent(), null);
 });
 
-tape("mercator.clipExtent(extent) intersects the specified clip extent with the automatic clip extent", function(test) {
-  var projection = d3.geoMercator().translate([0, 0]).scale(1).clipExtent([[-10, -10], [10, 10]]).precision(0);
-  test.pathEqual(d3.geoPath(projection)({type: "Sphere"}), "M3.141593,-10L3.141593,0L3.141593,10L3.141593,10L-3.141593,10L-3.141593,10L-3.141593,0L-3.141593,-10L-3.141593,-10L3.141593,-10Z");
-  test.deepEqual(projection.clipExtent(), [[-10, -10], [10, 10]]);
-  test.end();
+it("mercator.clipExtent(extent) intersects the specified clip extent with the automatic clip extent", () => {
+  const projection = d3.geoMercator().translate([0, 0]).scale(1).clipExtent([[-10, -10], [10, 10]]).precision(0);
+  assertPathEqual(d3.geoPath(projection)({type: "Sphere"}), "M3.141593,-10L3.141593,0L3.141593,10L3.141593,10L-3.141593,10L-3.141593,10L-3.141593,0L-3.141593,-10L-3.141593,-10L3.141593,-10Z");
+  assert.deepStrictEqual(projection.clipExtent(), [[-10, -10], [10, 10]]);
 });
 
-tape("mercator.clipExtent(extent).scale(scale) updates the intersected clip extent", function(test) {
-  var projection = d3.geoMercator().translate([0, 0]).clipExtent([[-10, -10], [10, 10]]).scale(1).precision(0);
-  test.pathEqual(d3.geoPath(projection)({type: "Sphere"}), "M3.141593,-10L3.141593,0L3.141593,10L3.141593,10L-3.141593,10L-3.141593,10L-3.141593,0L-3.141593,-10L-3.141593,-10L3.141593,-10Z");
-  test.deepEqual(projection.clipExtent(), [[-10, -10], [10, 10]]);
-  test.end();
+it("mercator.clipExtent(extent).scale(scale) updates the intersected clip extent", () => {
+  const projection = d3.geoMercator().translate([0, 0]).clipExtent([[-10, -10], [10, 10]]).scale(1).precision(0);
+  assertPathEqual(d3.geoPath(projection)({type: "Sphere"}), "M3.141593,-10L3.141593,0L3.141593,10L3.141593,10L-3.141593,10L-3.141593,10L-3.141593,0L-3.141593,-10L-3.141593,-10L3.141593,-10Z");
+  assert.deepStrictEqual(projection.clipExtent(), [[-10, -10], [10, 10]]);
 });
 
-tape("mercator.clipExtent(extent).translate(translate) updates the intersected clip extent", function(test) {
-  var projection = d3.geoMercator().scale(1).clipExtent([[-10, -10], [10, 10]]).translate([0, 0]).precision(0);
-  test.pathEqual(d3.geoPath(projection)({type: "Sphere"}), "M3.141593,-10L3.141593,0L3.141593,10L3.141593,10L-3.141593,10L-3.141593,10L-3.141593,0L-3.141593,-10L-3.141593,-10L3.141593,-10Z");
-  test.deepEqual(projection.clipExtent(), [[-10, -10], [10, 10]]);
-  test.end();
+it("mercator.clipExtent(extent).translate(translate) updates the intersected clip extent", () => {
+  const projection = d3.geoMercator().scale(1).clipExtent([[-10, -10], [10, 10]]).translate([0, 0]).precision(0);
+  assertPathEqual(d3.geoPath(projection)({type: "Sphere"}), "M3.141593,-10L3.141593,0L3.141593,10L3.141593,10L-3.141593,10L-3.141593,10L-3.141593,0L-3.141593,-10L-3.141593,-10L3.141593,-10Z");
+  assert.deepStrictEqual(projection.clipExtent(), [[-10, -10], [10, 10]]);
 });
 
-tape("mercator.rotate(…) does not affect the automatic clip extent", function(test) {
-  var projection = d3.geoMercator(), object = {
+it("mercator.rotate(…) does not affect the automatic clip extent", () => {
+  const projection = d3.geoMercator(), object = {
     type: "MultiPoint",
     coordinates: [
       [-82.35024908550241, 29.649391549778745],
@@ -49,10 +43,9 @@ tape("mercator.rotate(…) does not affect the automatic clip extent", function(
     ]
   };
   projection.fitExtent([[0, 0], [960, 600]], object);
-  test.deepEqual(projection.scale(), 20969742.365692537);
-  test.deepEqual(projection.translate(), [30139734.76760269, 11371473.949706702]);
+  assert.deepStrictEqual(projection.scale(), 20969742.365692537);
+  assert.deepStrictEqual(projection.translate(), [30139734.76760269, 11371473.949706702]);
   projection.rotate([0, 95]).fitExtent([[0, 0], [960, 600]], object);
-  test.deepEqual(projection.scale(), 35781690.650920525);
-  test.deepEqual(projection.translate(), [75115911.95344563, 2586046.4116968135]);
-  test.end();
+  assert.deepStrictEqual(projection.scale(), 35781690.650920525);
+  assert.deepStrictEqual(projection.translate(), [75115911.95344563, 2586046.4116968135]);
 });
