@@ -1,40 +1,36 @@
-var tape = require("tape"),
-    d3 = require("../../");
+import assert from "assert";
+import * as d3 from "../../src/index.js";
+import {assertProjectionEqual} from "./asserts.js";
+import {assertInDelta} from "../asserts.js";
 
-require("../inDelta");
-require("./projectionEqual");
-
-tape("projection.reflectX(…) defaults to false", function(test) {
-  var projection = d3.geoGnomonic().scale(1).translate([0, 0]);
-  test.equal(projection.reflectX(), false);
-  test.equal(projection.reflectY(), false);
-  test.projectionEqual(projection, [0, 0], [0, 0]);
-  test.projectionEqual(projection, [10, 0], [0.17632698070846498, 0]);
-  test.projectionEqual(projection, [0, 10], [0, -0.17632698070846498]);
-  test.end();
+it("projection.reflectX(…) defaults to false", () => {
+  const projection = d3.geoGnomonic().scale(1).translate([0, 0]);
+  assert.strictEqual(projection.reflectX(), false);
+  assert.strictEqual(projection.reflectY(), false);
+  assertProjectionEqual(projection, [0, 0], [0, 0]);
+  assertProjectionEqual(projection, [10, 0], [0.17632698070846498, 0]);
+  assertProjectionEqual(projection, [0, 10], [0, -0.17632698070846498]);
 });
 
-tape("projection.reflectX(…) mirrors x after projecting", function(test) {
-  var projection = d3.geoGnomonic().scale(1).translate([0, 0]).reflectX(true);
-  test.equal(projection.reflectX(), true);
-  test.projectionEqual(projection, [0, 0], [0, 0]);
-  test.projectionEqual(projection, [10, 0], [-0.17632698070846498, 0]);
-  test.projectionEqual(projection, [0, 10], [0, -0.17632698070846498]);
+it("projection.reflectX(…) mirrors x after projecting", () => {
+  const projection = d3.geoGnomonic().scale(1).translate([0, 0]).reflectX(true);
+  assert.strictEqual(projection.reflectX(), true);
+  assertProjectionEqual(projection, [0, 0], [0, 0]);
+  assertProjectionEqual(projection, [10, 0], [-0.17632698070846498, 0]);
+  assertProjectionEqual(projection, [0, 10], [0, -0.17632698070846498]);
   projection.reflectX(false).reflectY(true);
-  test.equal(projection.reflectX(), false);
-  test.equal(projection.reflectY(), true);
-  test.projectionEqual(projection, [0, 0], [0, 0]);
-  test.projectionEqual(projection, [10, 0], [0.17632698070846498, 0]);
-  test.projectionEqual(projection, [0, 10], [0, 0.17632698070846498]);
-  test.end();
+  assert.strictEqual(projection.reflectX(), false);
+  assert.strictEqual(projection.reflectY(), true);
+  assertProjectionEqual(projection, [0, 0], [0, 0]);
+  assertProjectionEqual(projection, [10, 0], [0.17632698070846498, 0]);
+  assertProjectionEqual(projection, [0, 10], [0, 0.17632698070846498]);
 });
 
-tape("projection.reflectX(…) works with projection.angle()", function(test) {
-  var projection = d3.geoMercator().scale(1).translate([10, 20]).reflectX(true).angle(45);
-  test.equal(projection.reflectX(), true);
-  test.inDelta(projection.angle(), 45);
-  test.projectionEqual(projection, [0, 0], [10, 20]);
-  test.projectionEqual(projection, [10, 0], [9.87658658, 20.12341341]);
-  test.projectionEqual(projection, [0, 10], [9.87595521, 19.87595521]);
-  test.end();
+it("projection.reflectX(…) works with projection.angle()", () => {
+  const projection = d3.geoMercator().scale(1).translate([10, 20]).reflectX(true).angle(45);
+  assert.strictEqual(projection.reflectX(), true);
+  assertInDelta(projection.angle(), 45);
+  assertProjectionEqual(projection, [0, 0], [10, 20]);
+  assertProjectionEqual(projection, [10, 0], [9.87658658, 20.12341341]);
+  assertProjectionEqual(projection, [0, 10], [9.87595521, 19.87595521]);
 });
